+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Changelog
 **1.1.0**
 - Improved categories DB storage adding categories_parent_categories table Which gives the possibility to store the same categories for many different parent categories
+- Added unique key constraints to the categories, attributes, attributes_values tables
 
 **1.0.0**
 - Created Migration script and sql files